Accepted empty char literals in the lexer (#97)
The one thing I know is that EOF_CHAR is the wrong error here, because
it's not actually an EOF (it's a grammar error). However, given that
we don't look at the contents of the char literal at all here (and
basically accept arbitrary string), I think this error should happen
at a higher level, together will all other such errors. As a result,
I propose removing it here.