Copy over high level structure of flisp-based parser
These function definitions and comments were automatically reformatted
into Julia code with the help of the script in
tools/flisp_defines_to_julia.jl with very light editing.

The idea is to try mirroring the structure and naming of the flisp-based
parser as closely as possible so that
* The code is mutually understandable
* The edge cases of Julia syntax are easy to capture by reusing the same
  recursive parser.

Perhaps this won't actually mesh well with the need to produce a
lossless syntax tree, but let's try it.