Remove use of K"VarIdentifier" from parser
Removing this allows us to remove the VAR_IDENTIFIER/VarIdentifier token
kind, the VAR_EOF token error and the derived VarMacroName kind. The
small cost for removing all this is the need to remap the kind of the
raw string following `var` into an identifier in the parser.

Also fix parsing of var followed by opening brackets, as in

    var"x"(