Tokenize: Always treat .&& and .|| as tokens
In https://github.com/JuliaLang/julia/pull/39594 it was considered
non-breaking enough to change the tokenization of ".&&" from `.&`,`&` to
`.&&`.

For JuliaSyntax it seems simplest to just use the newer tokenization
for this and detect the use of the syntax in the parser, emitting an
error on older versions. This leads to the most neat predictable parser
errors.