Process triple quoted string trivia during parsing
Triple quoted strings are de-indented with fairly complicated rules
* Based on string content
* The position of interpolations across multiple tokens

Because indentation isn't part of the string data it should also ideally
be excluded from the string content within the green tree. That is, it
should be treated as separate whitespace trivia tokens. With this
separation things like formatting should be much easier. The same
reasoning goes for escaping newlines and following whitespace with
backslashes in normal strings.

The downside of detecting string trivia during parsing is that string
content is split over several tokens. Here we wrap these in the K"string"
kind (as is already used for interpolations). The individual chunks can
then be reassembled during Expr construction. A possible alternative
might be to reuse the K"String" and K"CmdString" kinds for groups of
string chunks (without interpolation).