Separate token output from internal nodes in ParseStream
This data rearrangement gives a cleaner separation between tokens (which
keep track of bytes in the source text) vs internal tree nodes (which
keep track of which tokens they cover). As a result it reduces the size
of the intermediate data structures.

As part of rewriting build_tree to use the new data structures it's also
become much faster and building the green tree no longer dominates the
parsing time (probably due to fixing some type stability issues).