Cleanup: put next_byte in token rather than first_byte
While a bit counter-intuitive, this enables us to use an initial
sentinel token for recording the first byte of the first real token
which removes a bunch of special case hacks for computing the last byte
of the current output token.

Also return only SyntaxHead from peek_token() - the byte range is never
needed.