Always encapsulate strings in a string node
This change ensures that strings are always encapsulated within a
`K"string"` internal node of the parse tree, giving a place to include
the delimiters and unifying interpolated strings / strings with internal
whitespace with plain string literals.

Also wrap backtick delimited strings with a `K"cmdstring"` head for
the same reasons.