Parse keyword args with `=` head rather than `kw` (#103)
For Expr there's various cases where = is parsed into a kw head, but
this is inconsistent especially when named tuples come into play. That
is, the `=` parses to different heads:
* as `=`  in `(a=1, b=2)`
* as `kw` in `f(a=1, b=2)`

This causes extra complexity in the parser, and is visually confusing
for macro writers. Instead, this change always parses `=` to the K"="
head, converting to the :kw head when lowering to Expr.

This all seems to work fairly well, with one fairly obscure but awkward
exception - in infix notation such as `(x = 1) != y` the equality is
assignment, not a keyword argument.