Rework JuliaSyntax.parse() public API
`parse()` and `parseall()` were generally pretty inconvenient to use.

This change reworks what I had called `parseall()` to be more similar to
`Meta.parse()` and adds `parseall()` and `parseatom()` in analogy to the
`Base.Meta` versions of these functions.

The lower level function `parse!()` is provided to work with
`ParseStream` for cases where more control is required.