Revert "Parse keyword args with `=` head rather than `kw` (#103)"
It turns out that there's some problematic syntactic edge cases which
make this approach almost-but-not-quite work. In particular the
difference between

    -(a=1)
    -(a=1,)

Is the difference between assignment (first case) vs a keyword argument
(second case).

There's some potential ways to fix this without reverting (for example,
adding a new flag or syntax head for an "operator call" vs a normal
call). But for now reverting would seem simplest.