Report bad macro names more clearly (#147)
This change prevents a crash for various cases of bad syntax involving
macro names. In particular things like `@[x] a` and `@A.x a`.

The latter case where the @ is separated from the actual macro name is a
pain to deal with and is only approximately correct in this PR. But I'm
not sure why we even support this syntax and maybe we could
warn about it or deprecate it. So unclear whether going to greater
lengths emitting K"TOMBSTONE" to be precise about error reporting in
that case is worth it.