Add PARENS_FLAG to tuple, block and macrocall
Several syntactic constructs can occur either with or without
parentheses and it can be useful to distinguish between these without
looking at the syntax trivia - particularly for code formatting, but
also for other reasons.

* Macro calls: `@x(a,b)` vs `@x a b`
* Blocks: `(a; b)` vs `begin a ; b end`
* Tuples: `a,b` vs `(a,b)` (for example, see #194)

Also modify the printing of head flags so that each flag is clearly
distinguished from every other flag with a `-`.