Better conversion to `Expr` with `parse(..., ignore_errors=true)`
This isn't a complete solution, but it improves the situation somewhat.
(It's unclear whether we should spend a lot of effort here as converting
to Expr is not a particularly expressive path for error propagation. We
probably need to plumb diagnostic support into the Julia runtime in a
more general way.)