Rename parsex -> parsestmt + rule=:toplevel -> rule=:all
Using the name parsestmt() is explicit about the fact that this parses a
single statement (not an "expression" which is very ambiguous given
every Julia construct is an expression). It's also consistent with the
naming rules of `parseall()` and `parseatom()`, which already appear in
Base as Meta.parseall and Meta.parseatom.

Using rule=:all in the `parse!` API is most consistent with the choices
made in the Core._parser interface and the naming `parseall()`.