Fix combined cartesian / flattened generators (#263)
The combined cartesian/flattened generator syntax is a very unusual use
case found in only one or two packages in the general registry. It looks
like this

    ((a,b,c,d) for a=as, b=bs for c=cs, d=ds)

Practically all such generators end up as a flat list (though ... should
they?) but the ordering of tuples (a,b,c,d) in the list depends on which
parts are cartesian (separated by commas) and which parts are flattened
(separated by for's).

Here we fix the `SyntaxNode` representation of these, preferring a single
`generator` head rather than combined `flatten` and `generator`, and
instead grouping the sets of cartesian iterations into
`cartesian_iterator` groups. This keeps the tree structure flat and
closer to the source text while also faithfully representing these
unusual forms.  For example, the above parses as

    (generator
      (tuple a b c d)
      (cartesian_iterator
        (= a as)
        (= b bs))
      (cartesian_iterator
        (= c cs)
        (= d ds)))

In addition, we also make use of the `cartesian_iterator` head in `for`
loops, replacing the `block` which was used previously (but is not
semantically or syntactically a block!). Thus `for i=is, j=js body end`
now parses as

    (for
      (cartesian_iterator
        (= i is)
        (= j js))
      body)

This also improves Expr conversion, removing some special cases which
were necessary when processing `block's` in that situation.