Remove K"parens" nodes from SyntaxNode tree
Possible fix for #239:

* Add a `head` field to `SyntaxNode` to separate it from `GreenNode`
* When a `K"parens"` node is encountered in the green tree, replace it with its child.
* However, have the replacement point to the parens `GreenNode` so we can still observe parentheses

This turns out to work, but is awkward.  It's especially awkward/surprising for `sourcetext`, as we have things like

```julia
julia> kind(parsestmt(SyntaxNode, "(10)"))
K"Integer"

julia> sourcetext(parsestmt(SyntaxNode, "(10)"))
"(10)"
```

We could fix `sourcetext()` to avoid this inconsistency at the cost of traversing the raw tree dynamically.  But then do we also need to change `range()`?