Fix line numbers with source fragments
When parsing source code fragments incrementally with

    * `Meta.parse(str, index)` or
    * `parsestmt(str, index)`

we must avoid scanning the rest of `str` for line numbers for
efficiency. In this mode, the user is expected to provide `first_line`
to "manually" specify which line number we're counting from.

Admittedly this is a bit clunky and should be integrated better with
SourceFile (which should also be renamed - see issue #190) but for now
seems to be the most consistent way to approach things here.