Handle macros which call macros more correctly
Macros expansion of macros which emit macros as part of their AST needs
care when that AST contains interpolations: We don't want the second
macro expansion to see the hygenic_scope expression. One way to do this
is to transform hygenic_scope back into `SyntaxLiteral` on the fly.

Also fix auto-escaping for expressions arising from the original scope
by adding the necessary number of Expr(:esc)