Fix crashes due to lexing ambiguity of string delimiters (#394)
There are some lexing ambituities in primes vs cmd delimiters. We break
these with a simple rule in the lexer but there's edge cases of invalid
or extremely strange syntax where this can be inconsistent with the
parser. The following were some such cases which caused an assertion
error in the parser.

    "var\"#\"``\$"
    "x in'``\$"

This change avoids crashing in those cases, emitting an error instead.

See also #25