Add modular operators +% -% >>% etc.
For the last ten years, there have been idle discussion about allowing a way for
users to opt-in for explicitly overflowing integer operations.
Julia's arithmetic operations do overflow by default, but it may in the future be
possible to run Julia in a mode where overflows are detected, for testing
purposes. If this is to be done, then there needs to be a way to signal that the
user explicitly wants overflow, which is important in some low-level operations,
such as hashing.
For bitshifting, Julias definition of e.g. >> checks for overflow, which leads
to suboptimal code. Analogous to the opt-in overflowing arithmetic ops like +%,
users ought to be able to opt-in to the more efficient overflowing shifts, which
in situations where bitshifts are a performance bottleneck.