Replace K"true" and K"false" with K"Bool" (#488)
Use a single `K"Bool"` Kind for booleans. This is both more convenient
and more consistent with other literal kinds such as K"Integer" which
group all integers under a single kind.

Replace the use of the invisible `K"false"` token in catch blocks with a
new kind `K"Placeholder"` - this removes the last of the invisible
tokens, other than error tokens! K"Placeholder" will also be really
useful for JuliaLowering as a kind for all-underscore identifiers.