Disallow trailing junk in generators
Disallow generator syntax with trailing non-delimited expressions

    (y for x in xs a)
    (y for x in xs a, b)

Allow parameter syntax such as

    f(y for x in xs; a)

because the flisp parser allowed this it's used in Base in at least one
location.

Disallow the almost-equivalent block syntax for now:

    (y for x in xs; a)

because the flisp parser disallowed it and users can always get the same
thing with some extra parentheses.