AST: Use a single kind `K"op="` for updating assignments (#530)
Make all updating assignment operators like `+=` be represented with a
single `K"op="` head, with the operator itself in infix position.
For example, `x += 1` is now parsed as

    [op=]
      x    :: Identifier
      +    :: Identifier
      y    :: Identifier

This greatly reduces the number of distinct forms here from a rather big
list (`$=` `%=` `&=` `*=` `+=` `-=` `//=` `/=` `<<=` `>>=` `>>>=` `\=`
`^=` `|=` `÷=` `⊻=`) and makes the operator itself appear in the AST as
kind `K"Identifier"`, as it should. It also makes it possible to add
further unicode updating operators while keeping the AST stable.

The need for this was highlighted when working on JuliaLowering. When
using `K"+="` as a head, one needs to look up the appropriate operator
from the list of updating operators or use string munging on the Kind
itself. This is quite awkward especially as it needs special rules for
inferring the macro scope of the `+` identifier. In addition, having a
single head for this form means update operator semantics only need to
be dealt with in one place.