kinds: rm unused kind K"inert"
`inert` is mostly used in lowering. As far as I can tell, there
is precisely one place in the flisp parser that produces it:

https://github.com/JuliaLang/julia/blob/d6b3669621ceb18aea693d8544b2c38870d289ad/src/julia-parser.scm#L1259

However, we do not insert and `inert` there in JuliaSyntax, instead preferring
to automatically quote the rhs of a `.` head in expr conversion:

https://github.com/JuliaLang/JuliaSyntax.jl/blob/eceaa39706c6ecc3ef496613096216625dba558e/src/expr.jl#L310

As a result, (and as code coverage complains),  the syntax head is
unused and should be removed here. However, it should probably
then be added back in JuliaLowering.