organization: Split into logically distinct subpieces
A I mentioned in #560, and as contemplated in #536, I'd like to try re-using
JuliaParser infrastructure to replace parsers I've written for some other languages.
This takes the first step to do so by moving various files into directories depending
on whether they are language-dependent or not. Right now there is still some coupling
and of course, there are no actual abstractions between these pieces. The idea would
be to intrduce those over time. For now, if we put in this refactoring, the way to
use this would be to copy the appropriate pieces (at least `core/`) into your
downstream parser and then rewrite it to those APIs. I'm planning to do that with
a parser or two to see if I hit any big API issues and see what it would take
to actually make the re-use happen.

- core: Core functionality for parsing
- julia: Core functionality for parsing *julia*
- integration: Integration code to use as the parser for base
- porcelain: Other syntax tree types for external users of the package

The `integration` and `porcelain` components should not depend on each
other. Otherwise it's layered as expected. This is just the reorganization.
Additional work is required to actually spearate the abstractions.