Use implicit tree for error search (#562)
This builds on top of #560 and replaces the use of `SyntaxNode`
in hooks.jl by the new lower-level cursor APIs. This avoid allocating
two completely separate representations of the syntax tree.

As a result, the end-to-end parse time for error-containing code is between 1.5x
(if the error is the first token) and 2x (if the error is the last
token) faster than current master. However, the main motivation here
is just to reduce coupling between the Expr-producing and SyntaxNode
producing parts of the code.