Remove DOTOP_FLAG from tokenizer and parser (#568)
This implements the first of a series of AST format changes described
in #567. In particular, this removes the DOTOP_FLAG. Currently, we already
do not emit the DOTOP_FLAG on terminals, they always get split into one
dot token and one identifier token (although we did set it on the intermediate
tokens that came out of the lexer). The only four kinds for which DOTOP_FLAG
was ever set in the final AST were `=`, `op=`, `&&` and `||`. This introduces
separate head kinds for each of these (similar to how there are already
separate head calls for `dotcall` and `dot). Otherwise the AST structure
should be unchanged.