Abbreviated non-MIME `show()` for `GreenNode` (#581)
Non-MIME `show()` for `GreenNode` was using the default implementation
which is extremely verbose. Here we print it as an S-expression instead.
I've made a choice to show absolute position within the parent node
rather than node span so that it directly relates to indices in the
string it was derived from.