`_register_kinds!`: prevent unintentional closure capture, boxing
The variable `i` was unintentionally shared between the generator
closure in `_register_kinds!` and another part of the body of
`_register_kinds!`. Thus `i` was boxed, causing trouble for inference.

Fix this by moving the part of `_register_kinds!` with `i` to a new
function.

Fixing this should make the sysimage more resistant to invalidation,
once the change propagates to Julia itself.